Put the QR alignment patterns where the spec says
The phone pairing QR could not be scanned. The alignment-pattern spacing had
one division too many:

    Math.ceil((version * 4 + 4) / (numAlign * 2 - 2) / 2) * 2

Versions 1-6 have only the two outer centres, so step is never used and those
symbols were fine. From version 7 there is a middle centre, and the halved
step put it at 30 instead of 22. A misplaced 5x5 pattern corrupts the function
grid a decoder locks onto and overwrites data modules, so the symbol reads as
nothing at all.

The pairing URL is 140 characters, which lands on version 7 -- exactly the
first broken version. A sweep confirms the boundary: 134 chars decodes, 135
does not.

The encoder moves to lib/qr.ts so it can be tested, leaving phone.tsx with the
component. The test compares whole matrices against fixtures from a reference
implementation for versions 1, 7 and 11, and reverting the one character fails
the v7 and v11 cases.
