+++ desktop/src/lib/qr.test.ts
@@ -1,0 +1,209 @@
+import { describe, expect, it } from 'vitest'
+import { encodeQr } from './qr'
+
+/*
+ * Expected matrices come from Nayuki's `qrcodegen` 1.8.0 (byte mode, error-correction level L,
+ * `boostecl=False` — the same parameters this encoder uses), cross-checked by decoding the
+ * rendered symbols with zxing-cpp. Rows read left to right, '1' is a dark module.
+ * Regenerate with:
+ *   QrCode.encode_segments([QrSegment.make_bytes(text.encode())], QrCode.Ecc.LOW, boostecl=False)
+ */
+
+const PAIRING_URL = `https://thewh1teagle.github.io/vibe/phone/#${'a'.repeat(64)}:0123456789abcdef0123456789abcdef`
+const LONG_PAIRING_URL = `https://thewh1teagle.github.io/vibe/phone/#${'b'.repeat(64)}:${'0123456789abcdef'.repeat(2)}&extra=${'z'.repeat(150)}`
+
+const CASES: { name: string; text: string; version: number; rows: string[] }[] = [
+	{
+		name: 'short payload (version 1)',
+		text: 'vibe',
+		version: 1,
+		rows: [
+			'111111100010101111111',
+			'100000101010101000001',
+			'101110101011001011101',
+			'101110100000101011101',
+			'101110101111101011101',
+			'100000101110001000001',
+			'111111101010101111111',
+			'000000001000000000000',
+			'110100110011101110110',
+			'010011001101010101111',
+			'111101111101001111101',
+			'111111001110111111010',
+			'101110100111011100011',
+			'000000001010011011011',
+			'111111101011101011010',
+			'100000100110001000010',
+			'101110100100111000001',
+			'101110101000011101011',
+			'101110100011011101101',
+			'100000101001100101000',
+			'111111101010010101110',
+		],
+	},
+	{
+		name: 'a payload that fills version 1 exactly',
+		text: 'a'.repeat(17),
+		version: 1,
+		rows: [
+			'111111101100001111111',
+			'100000100100001000001',
+			'101110101111001011101',
+			'101110101110101011101',
+			'101110101001101011101',
+			'100000100111001000001',
+			'111111101010101111111',
+			'000000000000000000000',
+			'111100101010010011101',
+			'100011001010000110101',
+			'011101110010011100011',
+			'101010000111000111000',
+			'110110101010101100000',
+			'000000001001101011000',
+			'111111100101000111000',
+			'100000100101110001110',
+			'101110100101000001101',
+			'101110101100110000000',
+			'101110101011110001100',
+			'100000101010011100001',
+			'111111101000011010100',
+		],
+	},
+	{
+		name: 'the 140-character pairing URL (version 7)',
+		text: PAIRING_URL,
+		version: 7,
+		rows: [
+			'111111100110110100010100010111110100101111111',
+			'100000101111000000110111101010001001001000001',
+			'101110100001110101100101101010101101001011101',
+			'101110101011110111101010110101110001101011101',
+			'101110100010101000011111110100110011101011101',
+			'100000101101011000101000101111000100001000001',
+			'111111101010101010101010101010101010101111111',
+			'000000000110001100011000110111001010100000000',
+			'111110111010011101101111110001010100010101010',
+			'001000000001100010000101000010101001110000111',
+			'010010100010000110101000101010011111011001110',
+			'111001010111110011110001100010001101100111100',
+			'000000101001110111110110110100010000010000010',
+			'101001000110001000001100010010110001110000101',
+			'011000101110100000101011101001000111101100110',
+			'000101000010110101010101111111001000110111110',
+			'000111101010100001110010111001000000011001000',
+			'111001000111110100000101000101101101110100111',
+			'011011100001100000111000101111000111101111110',
+			'010011011010010101100000100010101100100111100',
+			'011111111011110111101111101100110000111110010',
+			'000110001111001000001000110110110001100010101',
+			'011010101100011001101010101001000111101011010',
+			'011110001100101100111000100111101010100010110',
+			'111111111001111101101111111000010111111111000',
+			'110110010000100010011110110101110000110000111',
+			'100010110011100110101000001100010110110001110',
+			'011001010010110011100101100010101101001101100',
+			'000010101110010111101101111101110000001110001',
+			'001101000111101000010010110000110000000000001',
+			'000101100101100001101101001001000110100100010',
+			'111100001010110101111000100010011001011000110',
+			'101001111110100001100000010001110010000110000',
+			'010101011011010100011111110011100000110000101',
+			'000010111000000000101000001010001110110001110',
+			'011110011111010101100001110010101101001101100',
+			'100110110010010111101111100100010000111110011',
+			'000000001011101000011000110100110000100011001',
+			'111111101001111001111010101001000110101011010',
+			'100000100100101100001000100110101111100010110',
+			'101110101101111100101111110100010000111110000',
+			'101110101100100010101101110001100000000010100',
+			'101110101011100110110010101000010011110001101',
+			'100000101000110011101100010010101101110011100',
+			'111111101000010111110001110101010000111010010',
+		],
+	},
+	{
+		name: 'a long pairing URL (version 11, multiple EC blocks)',
+		text: LONG_PAIRING_URL,
+		version: 11,
+		rows: [
+			'1111111010100111110011001101011101000111011100011001101111111',
+			'1000001011100110100100011100111011100101010100111101101000001',
+			'1011101011110010011011010011110011011100011111101011101011101',
+			'1011101010111100000010101101001001000110111001000110101011101',
+			'1011101000001101010010111001111111011110111101000011001011101',
+			'1000001011000100101101000000100011111000110100100010001000001',
+			'1111111010101010101010101010101010101010101010101010101111111',
+			'0000000000101011000001111011100011111010001011001011000000000',
+			'1100111000100011101111010001111111101101000110111111100101111',
+			'0010000101110001011001111110000001010010111101000111001000010',
+			'0011011111011011010011010001111101001111011000011111011111000',
+			'0100100110100011011111011011100010101010010111011001010110110',
+			'1110101101100010111001000010110100111100001010111100100010100',
+			'1000010111010001000011111100100000000010101101000011011100110',
+			'0010111100100001011010011000011100011110101010010011011011000',
+			'0110010101011101111011010110100011001111011110011011111110111',
+			'1000101100010111011001101010110100111001000010111100100011100',
+			'1000010101010010010010010100101001000010111101001011011100110',
+			'0010011111010110111011111011011000011111101011001110001011000',
+			'1010110101001111111011001010100001011010011111101010111010111',
+			'0011111101011111111101010011000010111111000110111101100011100',
+			'1011000100110011010010111000011101001110111101000011011101000',
+			'1100001010100001001011111001011100000111001001011111001001100',
+			'1000100001000101011001011100110011111010011011001001001010111',
+			'1101111001001001110011010110110110101101011010111011110111101',
+			'0001000100001011001110111010000001011010111011000011011100000',
+			'1010001100011011010101000000011101000111011100011111001011000',
+			'0100100101110111011001000010100011101010010111011001010110100',
+			'1000111111001010111001000010111110011100011010111100111110100',
+			'0000100011110001000011111101100010000011001101000011100010110',
+			'1101101011010111011011111101101010011110101010010010101011000',
+			'1111100010100111111010110000100011001001011110111011100010111',
+			'0000111110101100011000011011111110111111000011111100111111100',
+			'1011110001010001110011001101110101000011011101010011000010110',
+			'1110101110111011011011100011111110011011001011000111000011000',
+			'0001100011011100011011010010111101011010011011101010011010111',
+			'1001101111110111111101010011011100111011000010111100100001100',
+			'1010000101001111010010111001110101010010111101000010100011000',
+			'1101111110001111010011111111010110010011001001011111000011100',
+			'1000100111010001000000111011111001111010001111001100001010111',
+			'0011001111100001101110000001011000111101010010111110000101101',
+			'0101010100001101011001111111100111000010101111000011000010000',
+			'0000111010100011010011011001100011001111011010011010000011010',
+			'1111010101001111011111000010101101101010010111011000001000100',
+			'1001011110111010111001000010101101011100000010111100011000101',
+			'0001100001001001000011111101110110000010001101000011001110110',
+			'1000111111001001011010011101111110011110101010010010100111000',
+			'0111110010110011111011010100111101001011011111011010011000111',
+			'0110111000110111011001101011111100111111000011011100001001100',
+			'1111100111111110010010010101110101000010111101011011000010110',
+			'0011111010101110111011111011111110011010101010010111100011000',
+			'1110100111111001111011001010111101011000011011101010011010111',
+			'1111001110100011111101010011111110111011000010111100111111100',
+			'0000000011000011010010111001100011000010111101000010100010000',
+			'1111111001001011001010011000101010010011001001011111101011100',
+			'1000001010000101011000111101100011111010001011001101100010111',
+			'1011101011010011110011000110111110111101010010111011111111101',
+			'1011101000011001001111101010011101010010111111000011100000001',
+			'1011101001010101010101011001001011001111001100011010111101010',
+			'1000001010001001011001000011010111101010010111011001000010110',
+			'1111111011010100111001000011000110111100010010111100110100111',
+		],
+	},
+]
+
+describe('encodeQr', () => {
+	for (const testCase of CASES) {
+		// Only a module-for-module match proves a scanner can read it: a plausible-looking
+		// matrix with, say, misplaced alignment patterns decodes as nothing at all.
+		it(`matches the reference symbol for ${testCase.name}`, () => {
+			const matrix = encodeQr(testCase.text)
+			const size = testCase.version * 4 + 17
+			expect(matrix).toHaveLength(size)
+			expect(matrix.map((row) => row.map((dark) => (dark ? '1' : '0')).join(''))).toEqual(testCase.rows)
+		})
+	}
+
+	it('rejects a payload no version can hold', () => {
+		expect(() => encodeQr('a'.repeat(3000))).toThrow()
+	})
+})
